Add deep_copy() of data_device to copy_func() 29

+++ simulation/src/godunov_solver/Godunov.hpp
@@ -45,7 +45,7 @@ public:
     void nextIteration(Real dt) final;
     void prepareNextOutput(Real& dt) final;
     void pdiExposeData() final;
-    int deep(int*, int*);
+    int deep(double*, double*);
     bool finished() const final;
     void printMonitoring(double t_tot) const final;
     bool shouldPrintInformation() const final;
